# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-exponent",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `base ^ exp` (real exponentiation) by delegating to\n`srvcs-floatpower`.",
        "description": "This service owns the *control flow* but delegates the arithmetic to its\ndependency, exactly as specified:\n\n1. ask `srvcs-floatpower` for `result = base.powf(exp)`.\n\nIf the dependency is unreachable it reports itself degraded (`503`); if the\ndependency rejects the input it forwards the `422`. Validation propagates\nfrom the dependency — this orchestrator does not validate inputs itself.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExponentResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "base",
          "exp"
        ],
        "properties": {
          "base": {
            "type": "number",
            "format": "double"
          },
          "exp": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ExponentResponse": {
        "type": "object",
        "required": [
          "base",
          "exp",
          "result"
        ],
        "properties": {
          "base": {
            "type": "number",
            "format": "double"
          },
          "exp": {
            "type": "number",
            "format": "double"
          },
          "result": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}